fastapi-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
transformers==4.40.1
torch --index-url https://download.pytorch.org/whl/cpu
pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
ftfy
regex
